refactor(archive): move action events out of lib
