+++ devtools/build_topology_projection.py
@@ -91,8 +91,8 @@ LIB_PREFIX_TO_SUBPACKAGE = {
     "raw_payload": "lib/raw_payload/",
     "artifact_taxonomy_": "lib/artifact_taxonomy/",
     "artifact_taxonomy": "lib/artifact_taxonomy/",
-    "action_event_": "lib/action_event/",
-    "action_events": "lib/action_event/",
+    "action_event_": "archive/action_event/",
+    "action_events": "archive/action_event/",
     "message_": "lib/message/",
     "messages": "lib/message/",
     "conversation_": "lib/conversation/",
@@ -207,7 +207,7 @@ TARGET_TO_OWNER = [
     ("polylogue/lib/viewport/", "lib-viewport"),
     ("polylogue/lib/raw_payload/", "lib-raw-payload"),
     ("polylogue/lib/artifact_taxonomy/", "lib-artifact-taxonomy"),
-    ("polylogue/lib/action_event/", "lib-action-event"),
+    ("polylogue/archive/action_event/", "archive-action-event"),
     ("polylogue/lib/message/", "lib-message"),
     ("polylogue/lib/conversation/", "lib-conversation"),
     ("polylogue/lib/semantic/", "lib-semantic"),
