+++ polylogue/cli/query_stats.py
@@ -15,10 +15,10 @@ import click
 from polylogue.cli.query_feedback import emit_no_results
 
 if TYPE_CHECKING:
+    from polylogue.archive.action_event.action_events import ActionEvent
     from polylogue.archive.filter.filters import ConversationFilter
     from polylogue.archive.query.spec import ConversationQuerySpec
     from polylogue.cli.shared.types import AppEnv
-    from polylogue.lib.action_event.action_events import ActionEvent
     from polylogue.lib.models import Conversation, ConversationSummary
     from polylogue.protocols import ConversationArchiveStatsStore
