+++ polylogue/products/resume.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Protocol
 
 from pydantic import Field
 
-from polylogue.lib.action_event.action_events import build_tool_calls_from_content_blocks
+from polylogue.archive.action_event.action_events import build_tool_calls_from_content_blocks
 from polylogue.lib.conversation.models import Conversation
 from polylogue.products.archive import (
     ArchiveProductUnavailableError,
