+++ tests/unit/cli/test_query_support_runtime.py
@@ -9,12 +9,12 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from polylogue.archive.action_event.action_events import ActionEvent
 from polylogue.archive.query.spec import ConversationQuerySpec
 from polylogue.cli import query_output, query_semantic, query_stats
 from polylogue.cli.query_actions import apply_modifiers, apply_transform, delete_conversations, resolve_stream_target
 from polylogue.cli.query_contracts import QueryDeliveryTarget, QueryMutationSpec, QueryOutputSpec
 from polylogue.cli.shared.types import AppEnv
-from polylogue.lib.action_event.action_events import ActionEvent
 from polylogue.lib.models import ConversationSummary
 from polylogue.lib.roles import Role
 from polylogue.lib.search_hits import ConversationSearchHit
