+++ tests/unit/core/test_repo_identity.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-from polylogue.lib.action_event.action_events import ActionEvent
+from polylogue.archive.action_event.action_events import ActionEvent
 from polylogue.lib.conversation.attribution import extract_attribution, extract_attribution_from_action_events
 from polylogue.lib.message.messages import MessageCollection
 from polylogue.lib.models import Conversation, Message
